backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
Django==4.2
django-cors-headers==3.5.0
django-environ==0.11.2
djangorestframework==3.14.0
google-auth==2.23.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pytz==2023.3.post1
rsa==4.9
sqlparse==0.4.4
urllib3==1.26.16
